+++ packages/headless-astro/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @drupal-canvas/headless-astro
 
+## 0.5.3
+
+### Patch Changes
+
+- Updated dependencies [e1fae30]
+  - @drupal-canvas/headless@0.9.0
+
 ## 0.5.2
 
 ### Patch Changes